Move classes to a separate package
