+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/problems/ignored/CliIgnoredProblemLevelRemappingManager.kt
@@ -1,17 +1,15 @@
-package com.jetbrains.plugin.structure.intellij.problems
+package com.jetbrains.plugin.structure.intellij.problems.ignored
 
-import com.fasterxml.jackson.core.exc.StreamReadException
-import com.fasterxml.jackson.databind.DatabindException
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
 import com.jetbrains.plugin.structure.base.problems.PluginProblem
-import com.jetbrains.plugin.structure.base.problems.PluginProblems
 import com.jetbrains.plugin.structure.base.problems.isInstance
+import com.jetbrains.plugin.structure.intellij.problems.IgnoredLevel
+import com.jetbrains.plugin.structure.intellij.problems.LevelRemappingDefinition
+import com.jetbrains.plugin.structure.intellij.problems.LevelRemappingDefinitions
+import com.jetbrains.plugin.structure.intellij.problems.ProblemSolutionHintProvider
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.io.IOException
-import java.net.URL
 import kotlin.reflect.KClass
+import com.jetbrains.plugin.structure.intellij.problems.ProblemLevelRemappingManager as ProblemLevelRemappingManager1
 
 const val CLI_IGNORED = "cli-ignored"
 
@@ -19,7 +17,7 @@ typealias PluginProblemId = String
 
 private val LOG: Logger = LoggerFactory.getLogger(CliIgnoredProblemLevelRemappingManager::class.java)
 
-class CliIgnoredProblemLevelRemappingManager(ignoredProblems: List<PluginProblemId> = emptyList()) : ProblemLevelRemappingManager, ProblemSolutionHintProvider {
+class CliIgnoredProblemLevelRemappingManager(ignoredProblems: List<PluginProblemId> = emptyList()) : ProblemLevelRemappingManager1, ProblemSolutionHintProvider {
   private val ignoredProblemDefinitionLoader = CliIgnoredProblemDefinitionLoader()
 
   private val problemClasses: List<CliIgnoredPluginProblem> = ignoredProblemDefinitionLoader.load()
@@ -76,43 +74,3 @@ class CliIgnoredProblemLevelRemappingManager(ignoredProblems: List<PluginProblem
 
 }
 
-data class CliIgnoredPluginProblem(val id: String, val pluginProblemClassFqn: String, val since: String) {
-  val pluginProblemClass: KClass<*>?
-    get() {
-      return PluginProblems.resolveClass(pluginProblemClassFqn)
-    }
-}
-
-private const val CLI_IGNORED_PROBLEMS_FILE_NAME = "plugin-problems-cli-muteable.json"
-
-class CliIgnoredProblemDefinitionLoader(private val jsonUrl: URL) {
-  constructor() : this(
-    CliIgnoredProblemLevelRemappingManager::class.java.getResource(CLI_IGNORED_PROBLEMS_FILE_NAME)
-      ?: throw IOException("Definition for problems that can be ignored in the CLI switch " +
-        "cannot be found at <$CLI_IGNORED_PROBLEMS_FILE_NAME>")
-  )
-
-  private val json = ObjectMapper()
-
-  @Throws(IOException::class)
-  fun load(): List<CliIgnoredPluginProblem> {
-    return try {
-      val rawDefinitions: List<Map<String, String>> = json.readValue(jsonUrl)
-      rawDefinitions.map {
-        CliIgnoredPluginProblem(
-          it["id"] ?: throw IOException("Missing 'id' field in the definition of a CLI ignored problem"),
-          it["class"]
-            ?: throw IOException("Missing 'class' field in the definition of a CLI ignored problem"),
-          it["since"] ?: throw IOException("Missing 'since' field in the definition of a CLI ignored problem")
-        )
-      }
-    } catch (e: IOException) {
-      throw IOException("Cannot load CLI ignorable problems definitions from <$jsonUrl>", e)
-    } catch (e: StreamReadException) {
-      throw IOException("Cannot parse CLI ignorable problems definitions from JSON in <$jsonUrl>", e)
-    } catch (e: DatabindException) {
-      throw IOException("Cannot deserialize CLI ignorable problems definitions from JSON in <$jsonUrl>", e)
-    }
-  }
-
-}
