+++ intellij-plugin-structure/tests/src/test/kotlin/com/jetbrains/plugin/structure/intellij/problems/ignored/CliIgnoredProblemLevelRemappingManagerTest.kt
@@ -1,5 +1,9 @@
-package com.jetbrains.plugin.structure.intellij.problems
+package com.jetbrains.plugin.structure.intellij.problems.ignored
 
+import com.jetbrains.plugin.structure.intellij.problems.DuplicatedDependencyWarning
+import com.jetbrains.plugin.structure.intellij.problems.ForbiddenPluginIdPrefix
+import com.jetbrains.plugin.structure.intellij.problems.IgnoredLevel
+import com.jetbrains.plugin.structure.intellij.problems.emptyLevelRemapping
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Assert.assertNotNull
